+++ Plugins/private/Authentication/AuthenticationException.cpp
@@ -1,6 +1,0 @@
-#include "AuthenticationException.h"
-
-GetCredentialException::GetCredentialException(const char* message) : std::runtime_error(message)
-{
-	
-}