+++ Plugins/private/Authentication/AuthenticationException.h
@@ -1,8 +1,0 @@
-#pragma once
-#include <stdexcept>
-
-class GetCredentialException : public std::runtime_error
-{
-public:
-	GetCredentialException(const char* message);
-};