+++ Plugins/private/Servers/ServersPlugin.cpp
@@ -12,7 +12,7 @@ namespace Stormancer
 {
 	void ServerPlugin::registerClientDependencies(ContainerBuilder& builder)
 	{
-		builder.registerDependency<Servers_Impl, AuthenticationService>().as<Servers>().singleInstance();
+		builder.registerDependency<Servers_Impl, Users::UsersApi>().as<Servers>().singleInstance();
 	}
 
 	void ServerPlugin::registerSceneDependencies(ContainerBuilder& builder, std::shared_ptr<Scene> scene)