+++ Plugins/public/Helloworld/Hello.h
@@ -1,14 +1,14 @@
 #pragma once
 
 #include "stormancer/Tasks.h"
-#include "stormancer/ClientAPI.h"
+#include "Users/ClientAPI.hpp"
 
 namespace Helloworld
 {
 	class Hello: public Stormancer::ClientAPI<Hello>
 	{
 	public:
-		Hello(std::weak_ptr<Stormancer::AuthenticationService> auth);
+		Hello(std::weak_ptr<Stormancer::Users::UsersApi> users);
 		pplx::task<std::string> world(std::string name);
 	};
 }