+++ Changelog.rst
@@ -18,11 +18,13 @@ Added
 - ``Users.hpp``: Header-only replacement for the Authentication plugin (``AuthenticationService`` and the like).
 - ``Party.hpp``: Header-only replacement for the Party plugin.
 - Added test code for ``Users.hpp`` and ``Party.hpp``.
+- ``Steam.hpp``: Header-only Steam plugin.
 
 Changed
 *******
 
 - ``AuthenticationService::sendRequestToUser()``: Breaking change to the data contract, requires the latest server-side Users plugin.
+- All plugings use the ``Users`` plugin instead of the ``Authentication`` plugin.
 
 Fixed
 *****
@@ -30,3 +32,8 @@ Fixed
 - ``AuthenticationService::sendRequestToUser()``: OriginId now has a correct value on the recipient's side.
 - Server-to-client RPCs now complete properly on the server when the client completes them without sending any data. Previously, they would hang until the client disconnected from the scene.
 - Fixed rare unobserved exceptions in ``ApiClient`` and ``RPC`` (prevents crashes in some cases, especially that could be triggered by breakpoints).
+
+Removed
+*******
+
+- Deprecated ``Authentication`` plugin has been removed.